LayoutTests/http/wpt/web-animations/interfaces/KeyframeEffect/processing-a-keyframes-argument-002-expected.txt


FAIL easing values are parsed correctly when set on a property-indexed keyframe Type error
FAIL easing values are parsed correctly when using a keyframe sequence effect.getKeyframes is not a function. (In 'effect.getKeyframes()', 'effect.getKeyframes' is undefined)
PASS Invalid easing values are correctly rejected when set on a property-indexed keyframe 
PASS Invalid easing values are correctly rejected when using a keyframe sequence 
FAIL Errors from invalid easings on a property-indexed keyframe are thrown after reading all properties assert_equals: All properties were read before throwing the easing error expected 2 but got 0
FAIL Errors from invalid easings on a keyframe sequence are thrown after reading all properties assert_throws: function "() => {
    new KeyframeEffect(target, [ kf1, kf2 ]);
  }" did not throw
